fix(betterUptime): make better uptime on success
